fixing broken things from reverted branch
